refactor: Update ParseCurlCommand and Parser.Parse to accept optional variables
The ParseCurlCommand function in parse.go and the Parse method in parser.go have been modified to accept optional variables. The variables are now passed as variadic arguments, allowing for more flexibility in the function calls. If variables are provided, they will be assigned to the 'variables' variable. This change improves the extensibility and usability of the code.